# Julia remove globals