--- a/src/plugins/julia/templates/lr.template.jl
+++ b/src/plugins/julia/templates/lr.template.jl
@@ -28,11 +28,7 @@ using DataStructures
 
 # Constants and globals
 const EOF = "\$"
-yytext = ""
-yylength = 0
-__res = nothing
-__loc = nothing
-should_capture_locations = {{{CAPTURE_LOCATIONS}}}
+const should_capture_locations = {{{CAPTURE_LOCATIONS}}}
 
 # Types
 struct SyntaxError <: Exception
@@ -56,6 +52,13 @@ Base.@kwdef mutable struct StackEntry
     symbol
     semanticvalue
     loc
+end
+
+Base.@kwdef mutable struct ParserData
+    yytext
+    yylength = 0
+    __res = nothing
+    __loc = nothing
 end
 
 # --------------------------------------------------------------
@@ -113,20 +116,22 @@ function parseend(value) end
     onParseBegin - a function to call when parsing begins
     onParseEnd - a function to call when parsing ends, should accept as a single argument with the parsed value result
 =#
-# Q for Dmetry: ok to default tokenizer here? other implementations throw but I wanted to keep the parse interface to only the string required
-function parse(ss::AbstractString; tokenizerInitFunction::Function = inittokenizer, onparsebegin::Function = parsebegin, onparseend::Function = parseend)
+function parse(ss::AbstractString; tokenizerinitfunction::Function = inittokenizer, onparsebegin::Function = parsebegin, onparseend::Function = parseend)
     # constants inserted by the parser generator
     productions = {{{PRODUCTIONS}}} # [[1, 2, "handler1"], [3, 4, "handler2], ...] i.e. Vector{Vector{Union{Integer, String}}}
     table = {{{TABLE}}} # i.e. Dict{Int, String}
 
+    # initialize our parser data
+    parserdata = ParserData(yytext = "", yylength = 0, __res = nothing, __loc = nothing)
+
     # initialization and prep for parsing
     !isnothing(onparsebegin) && onparsebegin()
-    tokenizerData = tokenizerInitFunction(ss)
+    tokenizerData = tokenizerinitfunction(ss)
     stack = Stack{Union{StackEntry,Integer}}()
     push!(stack, 0)
 
     # begin parsing
-    token = getnexttoken!(tokenizerData)
+    token = getnexttoken!(parserdata, tokenizerData)
     shiftedtoken = nothing
     while hasmoretokens(tokenizerData) || !isempty(stack)
         # get a token and look it up in our parsing table
@@ -144,7 +149,7 @@ function parse(ss::AbstractString; tokenizerInitFunction::Function = inittokeniz
             push!(stack, StackEntry(symbol = token.type, semanticvalue = token.value, loc = yyloc(token)))
             push!(stack, tryparse(Int, SubString(entry, 2)))
             shiftedtoken = token
-            token = getnexttoken!(tokenizerData)
+            token = getnexttoken!(parserdata, tokenizerData)
 
             # found "reduce" instruction, which starts with r then has <production number> to reduce by - i.e. r2 means "reduce by production 2"
         elseif entry[1] == 'r'
@@ -175,8 +180,8 @@ function parse(ss::AbstractString; tokenizerInitFunction::Function = inittokeniz
             symboltoproducewith = production[1]
             reducestackentry = StackEntry(symbol = symboltoproducewith, semanticvalue = nothing, loc = nothing)
             if hassemanticaction
-                global yytext = isnothing(shiftedtoken) ? nothing : shiftedtoken.value
-                global yylength = isnothing(shiftedtoken) ? 0 : length(shiftedtoken.value)
+                parserdata.yytext = isnothing(shiftedtoken) ? nothing : shiftedtoken.value
+                parserdata.yylength = isnothing(shiftedtoken) ? 0 : length(shiftedtoken.value)
                 semanticactionhandler = getfield(SyntaxParser, Symbol(production[3]))
                 semanticactionargs = semanticvalueargs
                 if should_capture_locations
@@ -184,10 +189,10 @@ function parse(ss::AbstractString; tokenizerInitFunction::Function = inittokeniz
                 end
 
                 # call the handler the result is put in __res, which is accessed/assigned to by for example $$ = <something> in the grammar
-                semanticactionhandler(semanticactionargs...)
-                reducestackentry.semanticvalue = __res
+                semanticactionhandler(parserdata, semanticactionargs...)
+                reducestackentry.semanticvalue = parserdata.__res
                 if should_capture_locations
-                    reducestackentry.loc = __loc
+                    reducestackentry.loc = parserdata.__loc
                 end
             end
             push!(stack, reducestackentry)