Removed non-static globals and put them in a structure which is passed, so should not be multithreading safe. Also did some performance optimizations.

Results of optimizations are:

BEFORE:

julia> @benchmark SyntaxParser.parse("5 + 5")
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  18.917 μs …  2.400 ms  ┊ GC (min … max): 0.00% … 98.15%
 Time  (median):     20.500 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   22.406 μs ± 63.097 μs  ┊ GC (mean ± σ):  7.87% ±  2.77%

 Memory estimate: 37.61 KiB, allocs estimate: 200.

AFTER:

 julia> @benchmark SyntaxParser.parse("5 + 5")
BenchmarkTools.Trial: 10000 samples with 3 evaluations.
 Range (min … max):   8.361 μs … 874.167 μs  ┊ GC (min … max):  0.00% … 98.16%
 Time  (median):      9.833 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):   11.513 μs ±  35.497 μs  ┊ GC (mean ± σ):  13.35% ±  4.27%

 Memory estimate: 28.77 KiB, allocs estimate: 96.